Bugfix: script failing with 'Set: Entry, ":AppleLocale", Does Not Exist'
This is because the plist for the simulator global preferences is
missing a key that the script is trying to alter. I was playing it safe
originally, but after further testing, I can just force remove both keys
I'm altering and then add them rather than try to update them.