Fixed issue #14: Problem choosing sim device on some macs.
The behavior to choose sim device is different if you have extra iOS
SDKs installed. If you don't, then the hardware device menu only shows
the device families as selectable menu items. If you do, then the device
familes are submenus and iOS versions are choosable below that.

This change makes the script more robust so it can work with both
setups.