Revamp run_screenshooter.sh to pass device and language to instruments
With new changes to instruments in Xcode 5.1, we can now pass the
simulator type directly on the command line. Coupled with what I learned
from [Oleg's][o] blog post, we can do away with the extra scripts that
muddled with the simulator and preferences. Everything to change device
type and language is specified entirely on the command line!!

  [o]: http://oleb.net/blog/2014/02/nsuserdefaults-handling-default-values/