Correct model when capturing landscape
The script capture.js might choose the wrong model when in landscape mode. This is because the model is determined from the height of the image. In landscape mode the height will always be 320 on iPhone, so a 4" landscape image will be saved as 3.5". 
The fix takes the orientation into account when looking for the 4" model.